Smoother install for Linux
My first instinct was just to run Main.py. This returned two
problems.

First, Main.py complained that I was using Python 2 instead
of Python 3. I fixed this by adding the crunch-bang line to
the top of Main.py and pointing it to python3. Now you can
just run it.

Second, it complained that it could not import pygame. This
was easily fixed with a "pip3 import pygame", but the
requirement wasn't listed anywhere in the documentation as
far as I saw. I added a python_requirements.txt file which
lists this requirement.

I then added "Install" and "Run" sections to the readme. This
information is Linux specific, so you might want to add
Windows and OSX instructions separately.